# Conflicts:
#	VictoryCenter/VictoryCenter.DAL/Data/VictoryCenterDbContext.cs
#	VictoryCenter/VictoryCenter.DAL/Entities/Category.cs
#	VictoryCenter/VictoryCenter.DAL/Migrations/VictoryCenterDbContextModelSnapshot.cs

+++ VictoryCenter/VictoryCenter.BLL/Commands/Categories/DeleteCategory/DeleteCategoryCommand.cs
@@ -3,5 +3,5 @@ using MediatR;
 
 namespace VictoryCenter.BLL.Commands.Categories.DeleteCategory;
 
-public record DeleteCategoryCommand(int id)
-    : IRequest<Result<int>>;
+public record DeleteCategoryCommand(long id)
+    : IRequest<Result<long>>;
